+++ src/server/actions/reachouts.ts
@@ -1,6 +1,6 @@
 'use server';
 
-import { OpportunityEmail } from '@/components/resend-emails/recruiter-dev-enquiry';
+import { OpportunityEmail } from '@/components/emails/recruiter-dev-enquiry';
 import { env } from '@/env';
 import {
   flushServerEvents,
